fix(clippy): Remove unnecessary casts in chromadb stats()
Fixes clippy error: unnecessary_cast
- Remove 'as usize' casts on lines 363-364
- count is already usize type

+++ src/knowledge/backends/chromadb.rs
@@ -360,8 +360,8 @@ impl VectorBackend for ChromaDbBackend {
         // For now, return total count
         // TODO: Add type-specific counts by querying with metadata filters
         Ok(BackendStats {
-            total_entries: count as usize,
-            success_count: count as usize,
+            total_entries: count,
+            success_count: count,
             correction_count: 0,
         })
     }
